hpack: fix readme example
While at it, might as well pull in the only other PR at this time
from the upstream hpack repo.

    https://github.com/mlalic/hpack-rs/pull/4

It doesn't affect the build as it is just a change to the README.md
but some people filed and found the issue, and one went so far to
propose this fix. (I haven't tried the fix.)